// src/app/models/unidadedoc/GetEnderecoModel.ts
export class GetEnderecoModel{
  
    idEndereco: string = '';
    dataHoraCriacao: string = '';
    cep: string = '';
    logradouro: string = '';
    complemento: string = '';
    numero: string = '';
    bairro: string = '';
    localidade: string = '';
    uf: string = '';
    ibge: string = '';
    gia: string = '';
    ddd: string = '';
    siafi: string = '';
 }